+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/v6/relation/RelationUpsertV6IT.kt
@@ -22,13 +22,13 @@ package org.eclipse.tractusx.bpdm.gate.v6.relation
 import org.assertj.core.api.Assertions
 import org.eclipse.tractusx.bpdm.common.dto.PageDto
 import org.eclipse.tractusx.bpdm.gate.api.model.RelationType
+import org.eclipse.tractusx.bpdm.gate.v6.UnscheduledGateV6TestBase
 import org.eclipse.tractusx.bpdm.gate.api.v6.model.request.RelationPutEntryV6
-import org.eclipse.tractusx.bpdm.gate.v6.GateUnscheduledInitialStartV6Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import org.springframework.web.reactive.function.client.WebClientResponseException
 
-class RelationUpsertV6IT: GateUnscheduledInitialStartV6Test() {
+class RelationUpsertV6IT: UnscheduledGateV6TestBase() {
 
     /**
      * GIVEN two business partners