+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/v6/stats/CountBusinessPartnersPerStageV6IT.kt
@@ -21,10 +21,10 @@ package org.eclipse.tractusx.bpdm.gate.v6.stats
 
 import org.assertj.core.api.Assertions
 import org.eclipse.tractusx.bpdm.gate.api.model.response.StatsStagesResponse
-import org.eclipse.tractusx.bpdm.gate.v6.GateUnscheduledInitialStartV6Test
+import org.eclipse.tractusx.bpdm.gate.v6.UnscheduledGateV6TestBase
 import org.junit.jupiter.api.Test
 
-class CountBusinessPartnersPerStageV6IT: GateUnscheduledInitialStartV6Test() {
+class CountBusinessPartnersPerStageV6IT: UnscheduledGateV6TestBase() {
 
     /**
      * GIVEN no business partners shared