+++ bpdm-pool-api/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/api/model/request/SiteCreateRequestWithLegalAddressAsMain.kt
@@ -21,12 +21,19 @@ package org.eclipse.tractusx.bpdm.pool.api.model.request
 
 import org.eclipse.tractusx.bpdm.common.dto.IBaseSiteDto
 import org.eclipse.tractusx.bpdm.pool.api.model.ConfidenceCriteriaDto
+import org.eclipse.tractusx.bpdm.pool.api.model.SiteHeaderDto
+import org.eclipse.tractusx.bpdm.pool.api.model.SiteHeaderScriptVariantDto
 import org.eclipse.tractusx.bpdm.pool.api.model.SiteStateDto
 
 data class SiteCreateRequestWithLegalAddressAsMain(
     override val name: String,
     override val states: Collection<SiteStateDto>,
     override val confidenceCriteria: ConfidenceCriteriaDto,
-    val bpnLParent: String
-) : IBaseSiteDto
+    val bpnLParent: String,
+    val scriptVariants: List<SiteHeaderScriptVariantDto> = emptyList()
+) : IBaseSiteDto{
+    fun toHeader(): SiteHeaderDto{
+        return SiteHeaderDto(name, states, confidenceCriteria, scriptVariants)
+    }
+}