+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/RelationUpsertService.kt
@@ -129,7 +129,7 @@ class RelationUpsertService(
         val legalEntityRelationType: LegalEntityRelationType,
         val validityPeriods: Collection<RelationValidityPeriodDb>,
         val existingRelation: RelationDb?,
-        val reasonCode: ReasonCodeDb
+        val reasonCode: ReasonCodeDb?
     )
 
     data class TimePeriod(